Crate turbo_tasks_hash

Hashing and encoding functions for turbopack.

An example use of this module is hashing a file’s content for cache invalidation, and encoding the hash to an hexadecimal string for use in a file name.

Structs

Xxh3Hash64Hasher

Xxh3Hash64 hasher.

Traits

DeterministicHash

DeterministicHash is a custom trait that signals the implementor can safely be hashed in a replicatable way across platforms and process runs. Note that the default Hash trait used by Rust is not deterministic for our purposes.

DeterministicHasher

DeterministicHasher is a custom trait that signals the implementor can safely hash in a replicatable way across platforms and process runs. Note that the default Hasher trait used by Rust allows for non-deterministic hashing, so it is not suitable for our purposes.

Functions

encode_base16

Encodes an array of bytes as a base16 string.

encode_hex

Encodes a 64-bit unsigned integer into a hex string.

encode_hex_string

Encodes a byte slice into a hex string.

hash_md4

Hash some content with the MD4 cryptographic hash function.

hash_xxh3_hash64

Hash some content with the Xxh3Hash64 non-cryptographic hash function.

Derive Macros

DeterministicHash