Struct turbo_tasks_hash::Xxh3Hash64Hasher

pub struct Xxh3Hash64Hasher(_);

Xxh3Hash64 hasher.

Implementations

impl Xxh3Hash64Hasher

pub fn new() -> Self

Create a new hasher.

pub fn write_value<T: DeterministicHash>(&mut self, input: T)

Uses the DeterministicHash trait to to hash the input in a cross-platform way.

pub fn write_ref<T: DeterministicHash>(&mut self, input: &T)

Uses the DeterministicHash trait to to hash the input in a cross-platform way.

pub fn finish(&self) -> u64

Finish the hash computation and return the digest.