Trait turbo_tasks_hash::DeterministicHash

pub trait DeterministicHash {
    // Required method
    fn deterministic_hash<H: DeterministicHasher>(&self, state: &mut H);
}

DeterministicHash is a custom trait that signals the implementor can safely be hashed in a replicatable way across platforms and process runs. Note that the default Hash trait used by Rust is not deterministic for our purposes.

It’s very important that Vcs never implement this, since they cannot be deterministic. The value that they wrap, however, can implement the trait.

Required Methods

fn deterministic_hash<H: DeterministicHasher>(&self, state: &mut H)

Adds self’s bytes to the [Hasher] state, in a way that is replicatable on any platform or process run.